backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
PyJWT
psycopg2-binary
SQLAlchemy